requirements.txt
beautifulsoup4==4.12.3
fastapi==0.110.1
uvicorn==0.29.0
nltk==3.8.1
passlib==1.7.3
pydantic==2.6.4
python_jose==3.3.0
Requests==2.31.0
SQLAlchemy==2.0.28
python-multipart
psycopg2-binary
pydantic[email]